修改sonar扫描的bug

See merge request zeta/dev/metaspace/metaspace!1499

+++ webapp/src/main/java/io/zeta/metaspace/web/rest/MetaDataREST.java
@@ -945,7 +945,9 @@ public class MetaDataREST {
             os.close();
             xlsxFile.delete();
         } finally {
-            xlsxFile.deleteOnExit();
+            if (null != xlsxFile) {
+                xlsxFile.deleteOnExit();
+            }
         }
     }
